Update schema in migrations
* Use unsignedBigInteger for foreign keys to other unsignedBigInteger
* Add missing foregin key constraint to orders table
* Change orders.commission_percent from integer to decimal able to store 0.00 through 0.99
  This is so that we can use it to multiply against order totals without converting the unit.
  However, the unit comes from the remote order processing system as a string, eg. "22%",
  which will need to first be converted to a decimal, eg. "0.22"

+++ laravel/database/migrations/2021_04_19_003108_create_line_items_table.php
@@ -15,10 +15,10 @@ class CreateLineItemsTable extends Migration
     {
         Schema::create('line_items', function (Blueprint $table) {
             $table->id();
-            $table->integer('quote_id');
+            $table->unsignedBigInteger('quote_id');
             $table->string('description');
             $table->decimal('price', 9, 2);
-            $table->integer('quantity');
+            $table->unsignedInteger('quantity');
             $table->timestamps();
             $table->foreign('quote_id')->references('id')->on('quotes');
         });
